refact: Refactorizar los componentes de autenticación y  del acceso denegado.
