+++ nitro/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { AuthProvider } from "@/components/auth/AuthProvider";
+import { AuthProvider } from "@/modules/auth/components/AuthProvider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
